feat: remove N from Flip, Roll, Symmetrize, CausalIntegration
