+++ examples/plot_causalintegration.py
@@ -112,9 +112,7 @@ ot = 0
 t = np.arange(nt) * dt + ot
 x = np.outer(np.sin(t), np.ones(nx))
 
-Cop = pylops.CausalIntegration(
-    nt * nx, dims=(nt, nx), sampling=dt, dir=0, halfcurrent=True
-)
+Cop = pylops.CausalIntegration((nt, nx), sampling=dt, dir=0, halfcurrent=True)
 
 y = Cop * x.ravel()
 y = y.reshape(nt, nx)
