+++ pylops/basicoperators/Roll.py
@@ -1,6 +1,7 @@
 import numpy as np
 
 from pylops import LinearOperator
+from pylops.utils._internal import _value_or_list_like_to_array
 
 
 class Roll(LinearOperator):
@@ -11,11 +12,8 @@ class Roll(LinearOperator):
 
     Parameters
     ----------
-    N : :obj:`int`
-        Number of samples in model.
-    dims : :obj:`list`, optional
+    dims : :obj:`list` or :obj:`int`
         Number of samples for each dimension
-        (``None`` if only one dimension is available)
     dir : :obj:`int`, optional
         Direction along which rolling is applied.
     shift : :obj:`int`, optional
@@ -39,31 +37,21 @@ class Roll(LinearOperator):
 
     """
 
-    def __init__(self, N, dims=None, dir=0, shift=1, dtype="float64"):
-        self.N = N
+    def __init__(self, dims, dir=0, shift=1, dtype="float64"):
+        self.dims = _value_or_list_like_to_array(dims)
+        N = np.prod(self.dims)
         self.dir = dir
-        if dims is None:
-            self.dims = (self.N,)
-            self.reshape = False
-        else:
-            if np.prod(dims) != self.N:
-                raise ValueError("product of dims must equal N")
-            else:
-                self.dims = dims
-                self.reshape = True
         self.shift = shift
-        self.shape = (self.N, self.N)
+        self.shape = (N, N)
         self.dtype = np.dtype(dtype)
         self.explicit = False
 
     def _matvec(self, x):
-        if self.reshape:
-            x = np.reshape(x, self.dims)
+        x = np.reshape(x, self.dims)
         y = np.roll(x, shift=self.shift, axis=self.dir)
         return y.ravel()
 
     def _rmatvec(self, x):
-        if self.reshape:
-            x = np.reshape(x, self.dims)
+        x = np.reshape(x, self.dims)
         y = np.roll(x, shift=-self.shift, axis=self.dir)
         return y.ravel()
