+++ pylops/basicoperators/Symmetrize.py
@@ -1,6 +1,7 @@
 import numpy as np
 
 from pylops import LinearOperator
+from pylops.utils._internal import _value_or_list_like_to_array
 from pylops.utils.backend import get_array_module
 
 
@@ -11,11 +12,8 @@ class Symmetrize(LinearOperator):
 
     Parameters
     ----------
-    N : :obj:`int`
-        Number of samples in model. Symmetric data has  :math:`2N-1` samples
-    dims : :obj:`list`, optional
+    dims : :obj:`list` or :obj:`int`
         Number of samples for each dimension
-        (``None`` if only one dimension is available)
     dir : :obj:`int`, optional
         Direction along which symmetrization is applied
     dtype : :obj:`str`, optional
@@ -43,8 +41,8 @@ class Symmetrize(LinearOperator):
         x[N-1-i],& \text{otherwise}
         \end{cases}
 
-    for :math:`i=0,1,2,\ldots,2N-2`, where :math:`N` is the lenght of
-    the input model.
+    for :math:`i=0,1,2,\ldots,2N-2`, where :math:`N` is the dimension of the input
+    model.
 
     In adjoint mode, the Symmetrize operator assigns the sums of the elements
     in position :math:`N-1-i` and :math:`N-1+i` to position :math:`i` as follows:
@@ -57,21 +55,11 @@ class Symmetrize(LinearOperator):
     apart from the central sample where :math:`x[0] = y[N-1]`.
     """
 
-    def __init__(self, N, dims=None, dir=0, dtype="float64"):
-        self.N = N
+    def __init__(self, dims, dir=0, dtype="float64"):
+        self.dims = _value_or_list_like_to_array(dims)
         self.dir = dir
-        if dims is None:
-            self.dims = (self.N,)
-            self.dimsd = (self.N * 2 - 1,)
-            self.reshape = False
-        else:
-            if np.prod(dims) != self.N:
-                raise ValueError("product of dims must equal N")
-            else:
-                self.dims = dims
-                self.dimsd = list(dims)
-                self.dimsd[self.dir] = dims[self.dir] * 2 - 1
-                self.reshape = True
+        self.dimsd = self.dims.copy()
+        self.dimsd[self.dir] = self.dims[self.dir] * 2 - 1
         self.nsym = self.dims[self.dir]
         self.shape = (np.prod(self.dimsd), np.prod(self.dims))
         self.dtype = np.dtype(dtype)
@@ -80,8 +68,7 @@ class Symmetrize(LinearOperator):
     def _matvec(self, x):
         ncp = get_array_module(x)
         y = ncp.zeros(self.dimsd, dtype=self.dtype)
-        if self.reshape:
-            x = ncp.reshape(x, self.dims)
+        x = ncp.reshape(x, self.dims)
         if self.dir > 0:  # bring the dimension to symmetrize to first
             x = ncp.swapaxes(x, self.dir, 0)
             y = ncp.swapaxes(y, self.dir, 0)
@@ -89,20 +76,17 @@ class Symmetrize(LinearOperator):
         y[: self.nsym - 1] = x[-1:0:-1]
         if self.dir > 0:
             y = ncp.swapaxes(y, 0, self.dir)
-        if self.reshape:
-            y = y.ravel()
+        y = y.ravel()
         return y
 
     def _rmatvec(self, x):
         ncp = get_array_module(x)
-        if self.reshape:
-            x = ncp.reshape(x, self.dimsd)
+        x = ncp.reshape(x, self.dimsd)
         if self.dir > 0:  # bring the dimension to symmetrize to first
             x = ncp.swapaxes(x, self.dir, 0)
         y = x[self.nsym - 1 :].copy()
         y[1:] += x[self.nsym - 2 :: -1]
         if self.dir > 0:
             y = ncp.swapaxes(y, 0, self.dir)
-        if self.reshape:
-            y = y.ravel()
+        y = y.ravel()
         return y
