+++ pylops/signalprocessing/_BaseFFTs.py
@@ -6,6 +6,7 @@ import numpy as np
 from numpy.core.multiarray import normalize_axis_index
 
 from pylops import LinearOperator
+from pylops.utils._internal import _raise_on_wrong_dtype, _value_or_list_like_to_array
 from pylops.utils.backend import get_complex_dtype, get_real_dtype
 
 logging.basicConfig(format="%(levelname)s: %(message)s", level=logging.WARNING)
@@ -15,59 +16,6 @@ class _FFTNorms(Enum):
     ORTHO = auto()
     NONE = auto()
     ONE_OVER_N = auto()
-
-
-def _value_or_list_like_to_array(value_or_list_like, repeat=1):
-    """Convert an object which is either single value or a list-like to an array.
-
-    Parameters
-    ----------
-    value_or_list_like
-        Single value or list-like.
-    repeat : `obj`:`int`
-        Size of resulting array if value is passed. If list is passed, it is ignored.
-
-    Returns
-    -------
-    out : `obj`:`numpy.array`
-        When the input is a single value, returned an array with `repeat` samples
-        containing that value. When the input is a list-like object, converts it to an
-        array.
-
-    """
-    try:
-        len(value_or_list_like)
-        out = np.array(value_or_list_like)
-    except TypeError:
-        out = np.array([value_or_list_like] * repeat)
-    return out
-
-
-def _raise_on_wrong_dtype(arr, dtype, name):
-    """Raises an error if dtype of `arr` is not a subdtype of `dtype`.
-
-    Parameters
-    ----------
-    arr : `obj`:`numpy.array`
-        Array whose type will be checked
-    dtype : `obj`:`numpy.dtype`
-        Type which must be a supertype of `arr.dtype`.
-    name : `obj`:`str`
-        Name of parameter to issue error.
-
-    Raises
-    ------
-    TypeError
-        When `arr.dtype` is not a subdtype of `dtype`.
-
-    """
-    if not np.issubdtype(arr.dtype, dtype):
-        raise TypeError(
-            (
-                f"Wrong input type for `{name}`. "
-                f'Must be "{dtype}", but received to "{arr.dtype}".'
-            )
-        )
 
 
 class _BaseFFT(LinearOperator):
