+++ pylops/waveeqprocessing/marchenko.py
@@ -401,8 +401,7 @@ class Marchenko:
             dtype=self.dtype,
         )
         Rollop = Roll(
-            self.nt2 * self.ns,
-            dims=(self.nt2, self.ns),
+            (self.nt2, self.ns),
             dir=0,
             shift=-1,
             dtype=self.dtype,
@@ -611,8 +610,7 @@ class Marchenko:
             dtype=self.dtype,
         )
         Rollop = Roll(
-            self.ns * nvs * self.nt2,
-            dims=(self.nt2, self.ns, nvs),
+            (self.nt2, self.ns, nvs),
             dir=0,
             shift=-1,
             dtype=self.dtype,
